requirements.txt
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
PyJWT==1.7.1
pytz==2018.7
requests==2.20.1
uritemplate==3.0.0
urllib3==1.24.1
wincertstore==0.2
